MangaDex support
Automatically informs a user if there are new chapters to read on
ManagDex. This is done by the user setting their RSS URL which
is saved in the database. The latest chapter is saved for refernce and
and all new chapters found with each check is reported back to the
user in the same channel the RSS URL was set.

cog fix

run files moved and added for windows

Renamed and updated

Renamed for syntax hightlight.
Updated to create table for Mangadex RSS
handling.

All test files in one location

Print error to console rather than flood a channel.

MangaDex support finished but buggy.

Create LICENSE

all required packages

No longer shows last seen with updates.

Refactored code

Possible request error handled

Handled a possible error when requesting the RSS data and the
status code for the request was not 200, resulting in "list
index out of range".

Function was not placed correctly, causing error.

Changed if logic (bug fix)

Missing colon

Database settings updated.