Merged modules and and more error handling.
Removed the PSQL module and created a helper module. Added a new file
for parsing RSS feeds, moving out this function from both rss.py and
mangadex.py. 

rss.py had a bug where it did not save the latest post when an URL was
added, and when a new post was detected, the user was not updated
correctly.

The RSS parser will now try five times before failing.